refactor: 주문 완료시 데이터 리패칭 로직 헤더로 이동
